+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.3
 - Death Move Automation Included: Add Scar and Update HP/Stress
+- Auto disable system Automation if Core is selected.
 
 # 1.2.2
 - Fixed duplicated audio